requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
alembic
psycopg2-binary
python-dotenv
httpx
pydantic
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
google-auth